server/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
openai
pdfplumber
python-docx
Pillow
pytesseract
pypdfium2
